### 2025-11-16 #9427 @editorialbot
[PRE REVIEW]: DNA Barcode Validator - a python toolkit for validating DNA barcodes
**Submitting author:** @rvosa (<a href="http://orcid.org/0000-0001-9254-7318">Rutger Vos</a>)
**Repository:** https://github.com/naturalis/barcode_validator
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.9
