### 2025-11-18 01ab412
Add status badge to README
Added a status badge for the project.
