### 2025-12-02 c413886
set default min length for rbcL to 500 bp
