+++ barcode_validator/criteria.py
@@ -179,7 +179,7 @@ class RBCLCriteria(MarkerCriteria):
         super().__init__(config)
         self.marker_type = Marker.RBCL
         # Default criteria for rbcL (these are example values)
-        self.min_length = 200  # min length for BOLD
+        self.min_length = 500  # min length for BOLD
         self.max_ambiguities = 0  # Skims must have no ambiguities
         self.max_stop_codons = 0  # No stop codons allowed
 
