### 2025-12-11 70db142
Add galaxy_blast.database config option and create test_rbcL.py
Co-authored-by: rvosa <106490+rvosa@users.noreply.github.com>
