### 2025-12-11 83e0358
Fix galaxy_blast.py imports and add Galaxy credential checks to test_rbcL.py
Co-authored-by: rvosa <106490+rvosa@users.noreply.github.com>
