### 2025-12-11 4ac5f9f
Remove auto-generated _version.py from git tracking and add to .gitignore
Co-authored-by: rvosa <106490+rvosa@users.noreply.github.com>
