### 2026-01-22 b0d59c9
Revise publication date and enhance paper content
Updated the publication date and added sections on software design, research impact, and AI usage disclosure.
