### 2026-02-10 65b478d
Restructure paper/paper.md to match required JOSS section format
Co-authored-by: rvosa <106490+rvosa@users.noreply.github.com>
