+++ paper/paper.md
@@ -22,15 +22,13 @@ bibliography: paper.bib
 
 # Summary
 
-DNA barcoding has become a cornerstone technique in molecular biodiversity research, enabling rapid species identification and discovery through standardized genetic markers. The Barcode Validator is a Python toolkit designed to ensure the quality and accuracy of DNA barcode sequences before submission to public databases such as the Barcode of Life Data System (BOLD) and institutional repositories. The software performs both structural validation (assessing sequence quality, length, ambiguous bases, and marker-specific features like stop codons) and taxonomic validation (verifying specimen identifications through reverse taxonomy using BLAST-based identification services). Developed to support large-scale biodiversity genomics initiatives, particularly the Biodiversity Genomics Europe (BGE) and ARISE projects, the toolkit provides automated workflows for processing thousands of sequences with flexible configuration options and comprehensive reporting.
+DNA barcoding has become a cornerstone technique in molecular biodiversity research, enabling rapid species identification and discovery through standardized genetic markers. The Barcode Validator is a Python toolkit designed to ensure the quality and accuracy of DNA barcode sequences before submission to public databases such as the Barcode of Life Data System (BOLD) and institutional repositories. The software performs both structural validation (assessing sequence quality, length, ambiguous bases, and marker-specific features like stop codons) and taxonomic validation (verifying specimen identifications through reverse taxonomy using BLAST-based identification services). Developed to support large-scale biodiversity genomics initiatives, particularly the Biodiversity Genomics Europe (BGE) and ARISE projects, the toolkit provides automated workflows for processing thousands of sequences with flexible configuration options and comprehensive reporting. The source code is available on GitHub at https://github.com/naturalis/barcode_validator under the Apache License 2.0. The software can be installed via PyPI (`pip install barcode-validator`) or Bioconda (`conda install -c bioconda barcode-validator`), and a Galaxy tool wrapper is available in the Galaxy ToolShed for web-based access.
 
 # Statement of need
 
 DNA barcoding projects generate large volumes of sequence data that must meet stringent quality standards before deposition in public databases. Manual validation of sequences is time-consuming, error-prone, and impractical for projects processing hundreds or thousands of specimens. Furthermore, the increasing adoption of genome skimming and high-throughput sequencing technologies produces multiple assembly attempts per specimen, requiring intelligent selection of the best valid sequence among alternatives.
 
-Existing validation tools are typically limited in scope: quality control tools like FastQC [@Andrews2010] focus on raw read quality rather than assembled barcode sequences; taxonomic identification tools like BOLD's identification engine [@Ratnasingham2007] or standalone BLAST [@Altschul1990] provide identification but lack integration with quality metrics; and no comprehensive solution exists for the specific workflow requirements of modern barcoding projects that combine multiple assembly attempts with both structural and taxonomic validation.
-
-The Barcode Validator addresses these gaps by providing:
+The Barcode Validator addresses these needs by providing:
 
 1. **Integrated validation**: Combined structural and taxonomic validation in a single workflow, with support for marker-specific requirements (e.g., stop codon detection for protein-coding genes via translation, GC content assessment for non-coding markers).
 
@@ -44,7 +42,13 @@ The Barcode Validator addresses these gaps by providing:
 
 The software has been deployed in production workflows at Naturalis Biodiversity Center (the Netherlands) and the Natural History Museum (United Kingdom) for the BGE project, processing thousands of arthropod COI sequences from genome skimming experiments, and the ARISE project for the validation of thousands of freshly sequenced vertebrate and marine and terrestrial invertebrate specimens. Its design supports the quality assurance requirements of modern DNA barcoding initiatives while remaining flexible enough to accommodate diverse project-specific workflows.
 
-# Implementation
[…]
