### 2026-02-10 3179bdc
Condense availability info in Summary per code review feedback
Co-authored-by: rvosa <106490+rvosa@users.noreply.github.com>
