+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-DNA barcoding has become a cornerstone technique in molecular biodiversity research, enabling rapid species identification and discovery through standardized genetic markers. The Barcode Validator is a Python toolkit designed to ensure the quality and accuracy of DNA barcode sequences before submission to public databases such as the Barcode of Life Data System (BOLD) and institutional repositories. The software performs both structural validation (assessing sequence quality, length, ambiguous bases, and marker-specific features like stop codons) and taxonomic validation (verifying specimen identifications through reverse taxonomy using BLAST-based identification services). Developed to support large-scale biodiversity genomics initiatives, particularly the Biodiversity Genomics Europe (BGE) and ARISE projects, the toolkit provides automated workflows for processing thousands of sequences with flexible configuration options and comprehensive reporting. The source code is available on GitHub at https://github.com/naturalis/barcode_validator under the Apache License 2.0. The software can be installed via PyPI (`pip install barcode-validator`) or Bioconda (`conda install -c bioconda barcode-validator`), and a Galaxy tool wrapper is available in the Galaxy ToolShed for web-based access.
+DNA barcoding has become a cornerstone technique in molecular biodiversity research, enabling rapid species identification and discovery through standardized genetic markers. The Barcode Validator is a Python toolkit designed to ensure the quality and accuracy of DNA barcode sequences before submission to public databases such as the Barcode of Life Data System (BOLD) and institutional repositories. The software performs both structural validation (assessing sequence quality, length, ambiguous bases, and marker-specific features like stop codons) and taxonomic validation (verifying specimen identifications through reverse taxonomy using BLAST-based identification services). Developed to support large-scale biodiversity genomics initiatives, particularly the Biodiversity Genomics Europe (BGE) and ARISE projects, the toolkit provides automated workflows for processing thousands of sequences with flexible configuration options and comprehensive reporting. The source code is available on GitHub at https://github.com/naturalis/barcode_validator under the Apache License 2.0, with distribution via PyPI and Bioconda, and a Galaxy tool wrapper for web-based access.
 
 # Statement of need
 
