### 2026-02-10 #9427 @rvosa
Dear @arfon, I have reworked the manuscript to conform to the structure as provided in the example paper (https://joss.readthedocs.io/en/latest/example_paper.html). Please let me know if this is in order. Thanks!

P.S. Something I was a bit unsure of is whether the PDF production pipeline also includes mermaid plots. I assume by now we all know and love what they look like on GitHub but it's a bit unclear if it's worthwhile including them for submission to JOSS.
