### 2026-02-10 9d90870
Add author D.S.J. Groenenberg to paper.md
