### 2026-02-11 35c7765
Add authors and affiliations to paper.md
