### 2026-02-12 b86aeed
Add author Rutger A. Vos to the paper
Added Rutger A. Vos as an author with corresponding status.
