### 2026-02-26 51c3cfa
skip undefined ranks; tribe, subfamily
