### 2026-02-26 5af6849
revert to default table if phylum is None
