### 2026-03-21 #9427 @majensen
Hi @amoeba - any chance you have time for [this one](#9427)? Thanks for considering!
