### 2026-03-21 #9427 @majensen
Hey @svburke - do you have time to review [this interesting software](https://github.com/openjournals/joss-reviews/issues/9427#9427)? --@majensen
