### 2026-03-29 #9427 @svburke
Hi @majensen, I will not have the time for this review.

As a non-reviewer suggestion though, it might be worth making this package a bit more Windows friendly in the future. Both of the Bioconda packages, BLAST and HMMER, require either Mac or Linux.

https://anaconda.org/channels/bioconda/packages/blast/overview
https://anaconda.org/channels/bioconda/packages/hmmer/overview
