### 2026-04-21 #9427 @majensen
Hello @iimog - would you have time to review [this work](https://github.com/openjournals/joss-reviews/issues/9427#9427) for [JOSS](https://joss.theoj.org/). I think your expertise is right in line. I appreciate your consideration - @majensen
