### 2026-04-21 #9427 @majensen
Hi @MikeTrizna - would you be able to review [this work](#9427) on validating DNA barcodes for [JOSS](https://joss.theoj.org)? I could use your expertise - let me know if you have any questions -- @majensen
