### 2026-04-21 #9427 @majensen
Bonjour @Juke34 - would you have time to review [this work](#9427), a DNA barcode validation system, for [JOSS](https://joss.theoj.org)? I would really appreciate your expertise - thanks for considering! -- @majensen
