### 2026-05-22 686a80d
fix: correct GitLab → GitHub references in CONTRIBUTING.md
Fixes issue #75
