### 2026-05-23 7dfaa90
Fix typo: GitLab -> GitHub in CONTRIBUTING.md
