### 2026-05-27 15ec144
Revise acknowledgements section for clarity
Updated acknowledgements to reflect authorship and support.
