+++ requirements.txt
@@ -7,3 +7,4 @@ openpyxl==3.1.2
 xlrd==2.0.1
 numpy==1.23.5
 nbitk>=0.7.2
+pytest
