### 2026-07-22 4ce5930
docs: distinguish Galaxy tool wrapper from Galaxy BLAST backend
