### 2026-07-22 2b946f7
test: add shared Galaxy credential guard in tests/conftest.py
