### 2026-07-22 2c0e72f
test: guard Galaxy tests with shared credential fixture and marker
