### 2026-07-22 36aed74
Update test_bge.py
