### 2026-07-22 264c37d
build: register the 'galaxy' pytest marker
