### 2026-07-22 90e0f1f
docs: document which tests need a Galaxy account and how to skip them
