### 2026-07-22 c503e28
docs: flag Galaxy-only backend and document all taxon-validation methods
