### 2026-07-22 4ae1cb1
Update galaxy_blast.py
