### 2026-07-22 5c8d0a4
Update schema.yaml
