+++ barcode_validator/config/schema.yaml
@@ -171,7 +171,9 @@ galaxy_blast:
 
 # The Galaxy BLAST database to use for taxonomic validation. This is the name of the database as it appears in
 # the Galaxy BLAST service. Different databases are available for different markers. For example, for COI-5P, the
-# default is "BOLD species only no duplicates", and for rbcL, you might use "Genbank rbcL (2023-11-15)".
+# default is "BOLD species only no duplicates", and for rbcL, you might use "Genbank rbcL (2023-11-15)". These
+# databases are hosted alongside the BLASTN tool on galaxy.naturalis.nl, which is at present the only instance
+# against which `--taxon-validation method=galaxy` can be run.
   database:
     type: str
     default: "BOLD species only no duplicates"
