### 2026-07-22 6a3b65c
docs: explain Galaxy-dependent tests in contributor guidelines
