### 2026-07-22 915a3f5
paper: qualify the Galaxy backend's instance requirement
