+++ paper/paper.md
@@ -74,7 +74,7 @@ Barcode Validator is implemented in Python (3.9+) with a modular, extensible arc
 
 The central design decision was to separate validation logic (what measurements to collect) from validity adjudication (what thresholds constitute pass/fail). This separation enables the same codebase to serve diverse projects with different quality requirements—genome skimming workflows demanding zero ambiguous bases versus Sanger sequencing tolerating several—without code modifications. The Strategy pattern for validators and Factory pattern for services enable runtime selection of validation approaches and identification backends. While this introduces abstraction overhead, it proved essential for accommodating the consortium's heterogeneous infrastructure: some partners operate local BLAST databases, others rely on Galaxy web services, and still others use BOLD's identification API directly.
 
-The software integrates with established bioinformatics tools including BLAST+ [@Camacho2009] for sequence similarity searches, HMMER [@Eddy2011] for profile Hidden Markov Model-based alignment and codon phase detection, and Biopython [@Cock2009] for sequence manipulation and translation. External validation is performed through REST API calls to BOLD [@Ratnasingham2007] and Galaxy [@Galaxy2022] identification services.
+The software integrates with established bioinformatics tools including BLAST+ [@Camacho2009] for sequence similarity searches, HMMER [@Eddy2011] for profile Hidden Markov Model-based alignment and codon phase detection, and Biopython [@Cock2009] for sequence manipulation and translation. External validation is performed through REST API calls to BOLD [@Ratnasingham2007] and Galaxy [@Galaxy2022] identification services. The BOLD backends require only an internet connection, and the local BLAST backend a user-supplied database, whereas the Galaxy backend requires an instance hosting a specific BLASTN-and-taxonomy tool together with curated reference libraries; at the time of writing this is the Naturalis instance, and the documentation directs users without access to it to the alternative backends.
 
 Input data can be provided as FASTA files with optional CSV metadata and BOLD Excel spreadsheets containing specimen and taxonomic information. Validation results are output in both human-readable TSV format (with detailed pass/fail status for each validation criterion) and filtered FASTA format (containing only sequences meeting all validation requirements).
 
@@ -82,7 +82,7 @@ Input data can be provided as FASTA files with optional CSV metadata and BOLD Ex
 
 The Barcode Validator has demonstrated substantial realized impact through its deployment in the Biodiversity Genomics Europe (BGE) project. As documented in BGE Deliverable D8.4, the toolkit processed sequences from over 18,500 specimens across 68 taxonomic orders, enabling the submission of more than 47,000 validated DNA barcode sequences to BOLD and the European Nucleotide Archive by October 2025. The validation framework identified systematic issues including plate-swap errors that would otherwise have corrupted database submissions, and revealed taxonomic patterns in validation success rates ranging from 0% to 100% across orders—insights that directly informed protocol optimizations.
 
-The software is deployed in production workflows at Naturalis Biodiversity Center (Netherlands) and the Natural History Museum (United Kingdom), with the ARISE project using it for validation of freshly sequenced vertebrate and invertebrate specimens. Community readiness is evidenced by: distribution through PyPI and Bioconda channels; availability as a Galaxy tool wrapper enabling web-based access for non-technical users; comprehensive documentation including architecture diagrams and use-case examples; an Apache 2.0 license; and a public GitHub repository with contribution guidelines. The toolkit's analytical outputs informed the BGE consortium's understanding of genome skimming assembly parameter optimization, demonstrating that the combination of specific preprocessing steps (*fcleaner*=TRUE, *merge*=FALSE) with alignment thresholds (*r*=1.0, *s*=50) maximizes barcode recovery while maintaining stringent quality standards.
[…]
